+++ UntilTheLastColorStand/core/src/main/java/com/gameonjava/utlcs/backend/resources/MovementPoint.java
@@ -22,10 +22,10 @@ public class MovementPoint extends Resource {
     }
 
 
-    public void updateMovementPoint(int techPoints){
+    public void updateMovementPoint(double techPoints){
         /* Updates the movement point based on tech points, techPoints/10 is just an constant you can edit
         that 10 whatever you want it to be */
-        int additionalMovementPoint = techPoints / 10;
+        double additionalMovementPoint = techPoints / 10;
         this.addResource(additionalMovementPoint);
     }
